extract workaround for focus issue on macOS
