+++ app/src/main-process/main.ts
@@ -2,7 +2,6 @@ import '../lib/logging/main/install'
 
 import { app, Menu, MenuItem, ipcMain, BrowserWindow, shell } from 'electron'
 import * as Fs from 'fs'
-import * as Url from 'url'
 
 import { AppWindow } from './app-window'
 import { buildDefaultMenu, MenuEvent, findMenuItemByID } from './menu'
@@ -14,6 +13,7 @@ import { fatalError } from '../lib/fatal-error'
 import { IMenuItemState } from '../lib/menu-update'
 import { LogLevel } from '../lib/logging/log-level'
 import { log as writeLog } from './log'
+import { openPathSafe } from './shell'
 import { reportError } from './exception-reporting'
 import {
   enableSourceMaps,
@@ -314,12 +314,7 @@ app.on('ready', () => {
         }
 
         if (stats.isDirectory()) {
-          const fileURL = Url.format({
-            pathname: path,
-            protocol: 'file:',
-            slashes: true,
-          })
-          shell.openExternal(fileURL)
+          openPathSafe(path)
         } else {
           shell.showItemInFolder(path)
         }
